app/requirements.txt
python-gnupg==0.4.9
google-cloud-storage==2.4.0
google-cloud-secret-manager==2.12.0
